Reverts "Fix bugs with changed attributes tracking when transaction gets rollback"
We are reverting these commits, because there are other caveats related
to dirty attributes not being restored when a transaction is rollback.
For instance, nested transactions cannot proper restore the dirty
attributes state after a rollback.

At the moment, the dirty attributes are scoped by the transaction.
When we call `.save` on a record, the dirty attributes will be reset even
if the transaction gets rollback.

[related #13166]
[related #15018]
[related #15016]
[closes #15019]

This reverts commits
* bab48f0a3d53a08bc23ea0887219e8deb963c3d9
  * b0fa7cf3ff8432cc2eef3682b34763b7f8c93cc8.
* 73fb39b6faa9de593ae75ad4e3b8e065ea0e53af
  * 37c238927fbed059de3f26a90d8923fb377568a5.
  * 8d8d4f1560264cd1c74364d67fa0501f6dd2c4fa

Revert "Merge pull request #13166 from bogdan/transaction-magic"