Update some bad test cases for serialized columns
The first case was not testing what the issue mentioned actually was (A
subclass of a class with serialized attributes does not serialize when
they come from the database).

The second case was a bad coder. It would fail if the model was `dup`ed,
or if the the model was loaded from the database and then saved again.

The third case wasn't testing anything that wasn't covered by the second
(and was also a bad coder for the same reasons as the second).